chore(types): dedupe IssuerRecord, Allocation
