+++ packages/deploy-script-support/src/offer.js
@@ -6,7 +6,7 @@ import { AmountMath } from '@agoric/ertp/src/amountMath.js';
 
 /**
  * @import {Petname} from '@agoric/deploy-script-support/src/externalTypes.js';
- * @import {InvitationDetails, Proposal} from '@agoric/zoe';
+ * @import {InvitationDetails, Keyword, Proposal} from '@agoric/zoe';
  */
 
 /**
