+++ packages/zoe/src/internal-types.js
@@ -2,7 +2,7 @@
 /// <reference types="@agoric/ertp/exported" />
 
 /**
- * @import {Completion, ContractStartFn, PaymentPKeywordRecord, UserSeat} from '@agoric/zoe';
+ * @import {Allocation, AnyTerms, BrandKeywordRecord, Completion, ContractStartFn, Handle, InvitationHandle, IssuerKeywordRecord, Keyword, PaymentPKeywordRecord, UserSeat, ZoeIssuerRecord} from '@agoric/zoe';
  * @import {Pattern} from '@endo/patterns';
  */
 
@@ -133,7 +133,7 @@
  * @property {ZoeInstanceAdminMakeInvitation} makeInvitation
  * @property {<I extends Issuer>(issuerP: ERef<I>,
  *             keyword: Keyword
- *            ) => Promise<I extends Issuer<infer K, infer M> ? IssuerRecord<K, M> : never>} saveIssuer
+ *            ) => Promise<I extends Issuer<infer K, infer M> ? ZoeIssuerRecord<K, M> : never>} saveIssuer
  * @property {MakeZoeMint} makeZoeMint
  * @property {RegisterFeeMint} registerFeeMint
  * @property {MakeNoEscrowSeat} makeNoEscrowSeat
@@ -196,7 +196,7 @@
 /**
  * @template {AssetKind} [K=AssetKind]
  * @typedef {object} ZoeMint
- * @property {() => IssuerRecord<K>} getIssuerRecord
+ * @property {() => import('@agoric/ertp').IssuerRecord<K>} getIssuerRecord
  * @property {(totalToMint: Amount<K>) => void} mintAndEscrow
  * @property {(totalToBurn: Amount<K>) => void} withdrawAndBurn
  * Note that the burning is asynchronous, and so may not have happened by
@@ -261,7 +261,7 @@
  */
 
 /**
- * @typedef {Array<IssuerRecord>} IssuerRecords
+ * @typedef {Array<ZoeIssuerRecord>} IssuerRecords
  */
 
 /**
@@ -296,7 +296,7 @@
  * contract instance
  *
  * @param {Keyword} keyword
- * @param {IssuerRecord} issuerRecord
+ * @param {ZoeIssuerRecord} issuerRecord
  * @returns {void}
  */
