+++ packages/zoe/src/zoeService/internal-types.js
@@ -3,6 +3,7 @@
 
 /**
  * @import {FeeMintAccess, GetBrands, GetBundleIDFromInstallation, GetIssuers, InstallBundle, InstallBundleID, SourceBundle} from './types.js';
+ * @import {InstanceRecord} from './utils.js';
  */
 
 /**
