+++ packages/zoe/src/zoeService/types.ts
@@ -14,10 +14,10 @@ import type {
   Keyword,
   InvitationHandle,
   BrandKeywordRecord,
-  Allocation,
   Handle,
   IssuerKeywordRecord,
 } from '../types.js';
+import type { Allocation } from '../types-index.js';
 
 /** @see {@link https://github.com/sindresorhus/type-fest/blob/main/source/is-any.d.ts} */
 type IsAny<T> = 0 extends 1 & NoInfer<T> ? true : false;
